requirements.txt
chardet==5.2.0
fastapi==0.128.0
loguru==0.7.3
pydantic==2.12.5
python-dotenv==1.2.1
Requests==2.32.5
